Updated installer with exceptions and more tools Added pyrit and crunch tools Added check in shell script if the python script fails

+++ install.sh
@@ -1,16 +1,20 @@
 ./tools.py
-cd cowpatty-4.6/
-sudo make
-echo Cowpatty maked, copying links
-cp cowpatty /usr/bin/cowpatty
-cp genpmk /usr/bin/genpmk
-cd ..
-echo Removing temp files
-rm -r cowpatty-4.6
-rm cowpatty-4.6.tgz
-echo
-echo WiCC installed
-echo
-echo You can now execute it with: wicc
-echo
-echo You can view the help with options -h or --help : wicc --help
+if [ $? -eq 0 ]; then
+    cd cowpatty-4.6/
+    sudo make
+    echo Cowpatty maked, copying links
+    cp cowpatty /usr/bin/cowpatty
+    cp genpmk /usr/bin/genpmk
+    cd ..
+    echo Removing temp files
+    rm -r cowpatty-4.6
+    rm cowpatty-4.6.tgz
+    echo
+    echo WiCC installed
+    echo
+    echo You can now execute it with: wicc
+    echo
+    echo You can view the help with options -h or --help : wicc --help
+else
+    echo Failed installation
+fi
